2015-07-21 Audio Description for Media and Live Performance
Audio Description (AD) makes visual images accessible for people who are blind or have low vision. Using words that are succinct, vivid and imaginative, description translates the visual image into an aural form that is accessible to these individuals.

In the performing arts (theater, dance, opera), and media (television, movies and DVD), description inserts this narrative into the natural pauses in dialogue or between critical sound elements.

Join us for this informal introduction to the challenges and application of the art and science of this burgeoning accessibility field.